Set runners to latest
The mac release was hard-coded to a now deprecated version. For release it might make more sense to use the latest runner always, its maybe expected that development builds can be tested on specific agent versions for legacy reasons but release presumably should be kept aligned with the latest agents? Otherwise another build tool might be a benefit such as vagrant.